+++ check-mysql/check-mysql.go
@@ -5,7 +5,6 @@ import (
 	"os"
 	"strings"
 
-	"github.com/jessevdk/go-flags"
 	"github.com/mackerelio/checkers"
 	"github.com/ziutek/mymysql/mysql"
 	_ "github.com/ziutek/mymysql/native"
@@ -16,18 +15,6 @@ type mysqlSetting struct {
 	Port string `short:"p" long:"port" default:"3306" description:"Port"`
 	User string `short:"u" long:"user" default:"root" description:"Username"`
 	Pass string `short:"P" long:"password" default:"" description:"Password"`
-}
-
-var cOpts struct {
-	mysqlSetting
-	Crit int64  `short:"c" long:"critical" default:"250" description:"critical if the number of connection is over"`
-	Warn int64  `short:"w" long:"warning" default:"200" description:"warning if the number of connection is over"`
-}
-
-var rOpts struct {
-	mysqlSetting
-	Crit int64  `short:"c" long:"critical" default:"250" description:"critical if the seconds behind master is over"`
-	Warn int64  `short:"w" long:"warning" default:"200" description:"warning if the seconds behind master is over"`
 }
 
 var commands = map[string](func([]string) *checkers.Checker){
@@ -62,80 +49,4 @@ Subcommand:
 func newMySQL(m mysqlSetting) mysql.Conn {
 	target := fmt.Sprintf("%s:%s", m.Host, m.Port)
 	return mysql.New("tcp", "", target, m.User, m.Pass, "")
-}
-
-func checkConnection(args []string) *checkers.Checker {
-	psr := flags.NewParser(&cOpts, flags.Default)
-	psr.Usage = "connection [OPTIONS]"
-	_, err := psr.ParseArgs(args)
-	if err != nil {
-		os.Exit(1)
-	}
-	db := newMySQL(cOpts.mysqlSetting)
-	err = db.Connect()
-	if err != nil {
-		return checkers.Unknown("couldn't connect DB")
-	}
-	defer db.Close()
-
-	rows, res, err := db.Query("SHOW GLOBAL STATUS LIKE 'Threads_Connected'")
-	if err != nil {
-		return checkers.Unknown("couldn't execute query")
-	}
-
-	idxValue := res.Map("Value")
-	threadsConnected := rows[0].Int64(idxValue)
-
-	checkSt := checkers.OK
-	msg := fmt.Sprintf("%d connections", threadsConnected)
-	if threadsConnected > cOpts.Crit {
-		checkSt = checkers.CRITICAL
-	} else if threadsConnected > cOpts.Warn {
-		checkSt = checkers.WARNING
-	}
-	return checkers.NewChecker(checkSt, msg)
-}
-
-func checkReplication(args []string) *checkers.Checker {
-	psr := flags.NewParser(&rOpts, flags.Default)
-	psr.Usage = "replication [OPTIONS]"
-	_, err := psr.ParseArgs(args)
-	if err != nil {
-		os.Exit(1)
-	}
-	db := newMySQL(rOpts.mysqlSetting)
-	err = db.Connect()
-	if err != nil {
-		return checkers.Unknown("couldn't connect DB")
-	}
-	defer db.Close()
-
-	rows, res, err := db.Query("SHOW SLAVE STATUS")
-	if err != nil {
-		return checkers.Unknown("couldn't execute query")
-	}
-
-	if len(rows) == 0 {
-		return checkers.Ok("MySQL is not slave")
-	}
-
-	idxIoThreadRunning := res.Map("Slave_IO_Running")
-	idxSQLThreadRunning := res.Map("Slave_SQL_Running")
-	idxSecondsBehindMaster := res.Map("Seconds_Behind_Master")
-	ioThreadStatus := rows[0].Str(idxIoThreadRunning)
-	sqlThreadStatus := rows[0].Str(idxSQLThreadRunning)
-	secondsBehindMaster := rows[0].Int64(idxSecondsBehindMaster)
-
-	if ioThreadStatus == "No" || sqlThreadStatus == "No" {
-		return checkers.Critical("MySQL replication has been stopped")
-	}
-
-	checkSt := checkers.OK
-	msg := fmt.Sprintf("MySQL replication behind master %d seconds", secondsBehindMaster)
-	if secondsBehindMaster > rOpts.Crit {
-		checkSt = checkers.CRITICAL
-	} else if secondsBehindMaster > rOpts.Warn {
-		checkSt = checkers.WARNING
-	}
-	return checkers.NewChecker(checkSt, msg)
 }
